+++ crates/cairo-coverage/tests/helpers/test_project.rs
@@ -39,6 +39,7 @@ impl TestProject {
             .output()
     }
 
+    #[cfg(feature = "allows-excluding-macros")]
     pub fn run_without_genhtml(self) -> TestProjectOutput {
         self.generate_trace_files().run_coverage().output()
     }
@@ -130,6 +131,7 @@ impl TestProjectOutput {
         assert_eq!(content, expected);
     }
 
+    #[cfg(feature = "allows-excluding-macros")]
     pub fn assert_empty_output(self) {
         assert!(self.read_output().is_empty());
     }