+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+#### Changed
+- macros are now by default included in the coverage report. If you want to exclude them, use the `--include` without the
+  `macros` option (can also have empty value)
+
 ## [0.4.0] - 2025-01-03
 
 #### Added