+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Added
 - warning if the current scarb version doesn't support not including macros in the coverage report
+- `--unstable` flag to enable unstable features. Read more about it [here](./README.md#coverage-across-different-scarb-versions)
 
 #### Changed
 - macros are now by default included in the coverage report. If you want to exclude them, use the `--include` without the